long-hint had special YAML char, so failed tests
I included a colon (:) in the revised long-hint and apparently that's interpreted as a special YAML character. @michaelsneddon [why can't I invoke people with the @? Does that not work on GitHub?] let me know that the tests failed (yay for tests!) and suggested the fix of using the pipe and putting the text on the following line. Hopefully that will work.